+++ src/components/AuthoritiesSearchForm/AuthoritiesSearchForm.js
@@ -53,6 +53,7 @@ const AuthoritiesSearchForm = ({
 
   const {
     navigationSegmentValue,
+    filters,
     searchDropdownValue,
     searchInputValue,
     setSearchInputValue,
@@ -113,6 +114,9 @@ const AuthoritiesSearchForm = ({
     || !searchInputValue
     || isAuthoritiesLoading;
 
+  const isFiltersApplied = Object.values(filters).find(value => value.length > 0);
+  const isResetAllButtonDisabled = (!searchInputValue && !isFiltersApplied) || isAuthoritiesLoading;
+
   return (
     <form onSubmit={onSubmitSearch}>
       <FilterNavigation />
@@ -154,7 +158,7 @@ const AuthoritiesSearchForm = ({
                 buttonStyle="none"
                 id="clickable-reset-all"
                 fullWidth
-                disabled={!searchInputValue || isAuthoritiesLoading}
+                disabled={isResetAllButtonDisabled}
                 onClick={() => {
                   resetRows();
                   handleResetAll();