+++ src/constants/index.js
@@ -6,3 +6,5 @@ export * from './searchResultsListColumns';
 export * from './filterConfig';
 export * from './sortOrders';
 export * from './navigationSegments';
+export * from './subjectHeadingsValues';
+export * from './subjectHeadingsMap';