+++ src/constants/subjectHeadingsMap.js
@@ -1,0 +1,15 @@
+import { subjectHeadingsValues } from './subjectHeadingsValues';
+
+export const subjectHeadingsMap = {
+  [subjectHeadingsValues.LIBRARY_OF_CONGRESS]: 'a',
+  [subjectHeadingsValues.LC_FOR_CHILDRENS_LITERATURE]: 'b',
+  [subjectHeadingsValues.MEDICAL]: 'c',
+  [subjectHeadingsValues.NATIONAL_AGRICULTURE_LIBRARY]: 'd',
+  [subjectHeadingsValues.CANADIAN]: 'k',
+  [subjectHeadingsValues.NOT_APPLICABLE]: 'n',
+  [subjectHeadingsValues.ART_AND_ARCHITECTURE]: 'r',
+  [subjectHeadingsValues.SEARS_LIST]: 's',
+  [subjectHeadingsValues.REPERTOIRE_DE_VEDETTES_MATIERE]: 'v',
+  [subjectHeadingsValues.OTHER]: 'z',
+  [subjectHeadingsValues.NOT_SPECIFIED]: ' ',
+};