+++ src/constants/subjectHeadingsValues.js
@@ -1,0 +1,13 @@
+export const subjectHeadingsValues = {
+  LIBRARY_OF_CONGRESS: 'Library of Congress Subject Headings',
+  LC_FOR_CHILDRENS_LITERATURE: 'LC subject headings for children\'s literature',
+  MEDICAL: 'Medical Subject Headings',
+  NATIONAL_AGRICULTURE_LIBRARY: 'National Agricultural Library subject authority file',
+  CANADIAN: 'Canadian Subject Headings',
+  NOT_APPLICABLE: 'Not applicable',
+  ART_AND_ARCHITECTURE: 'Art and Architecture Thesaurus',
+  SEARS_LIST: 'Sears List of Subject Headings',
+  REPERTOIRE_DE_VEDETTES_MATIERE: 'Répertoire de vedettes-matière',
+  OTHER: 'Other',
+  NOT_SPECIFIED: 'Not specified',
+};